Refactor: QueryDsl Sort related common class name change
